requirements.txt
python-telegram-bot==20.5
python-docx==0.8.11
requests==2.31.0
PyMuPDF==1.24.11